Updated mean subtraction for image-classifier This change does a mean subtraction on float16 as against float32. I don't expect this to have any affect on accuracy, but converting the input image back and forth between float16 & float32 seemed like an overhead on the CPU.

+++ apps/image-classifier/image-classifier.py
@@ -54,11 +54,11 @@ img = skimage.transform.resize( img, IMAGE_DIM, preserve_range=True )
 img = img[:, :, ::-1]
 
 # Mean subtraction & scaling [A common technique used to center the data]
-img = img.astype( numpy.float32 )
+img = img.astype( numpy.float16 )
 img = ( img - IMAGE_MEAN ) * IMAGE_STDDEV
 
 # Load the image as a half-precision floating point array
-graph.LoadTensor( img.astype( numpy.float16 ), 'user object' )
+graph.LoadTensor( img, 'user object' )
 
 # ---- Step 4: Read & print inference results from the NCS -------------------
